Lates maintenance changes from Development branch

+++ implementation/org.gravity.tgg.modisco.uml/src/org/gravity/tgg/uml/UmlSecProcessor.java
@@ -9,9 +9,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.apache.log4j.Logger;
-import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.ecore.EObject;
-import org.eclipse.emf.ecore.EStructuralFeature;
 import org.eclipse.uml2.uml.Classifier;
 import org.eclipse.uml2.uml.Comment;
 import org.eclipse.uml2.uml.Element;
@@ -65,62 +63,35 @@ public class UmlSecProcessor {
 
 	public UmlSecProcessor(final Model model) {
 		this.model = model;
-
-		final String[] namespace = new String[] { "org", "gravity", "security", "annotations", "requirements" };
-		final LinkedList<Model> stack = new LinkedList<>();
-		stack.add(model);
-		Package reqPack = null;
-		while (!stack.isEmpty()) {
-			final Model next = stack.pop();
-			reqPack = getPackage(next, namespace);
-			if (reqPack != null) {
-				break;
-			}
-			if (next != null) {
-				for (final PackageableElement element : next.getPackagedElements()) {
-					if (element instanceof Model) {
-						stack.add((Model) element);
-					}
-				}
-			}
-		}
-		if (reqPack != null) {
-			this.iCritical = (Interface) reqPack.getPackagedElement(ANNOTATION_CRITICAL);
-			this.iHigh = (Interface) reqPack.getPackagedElement(ANNOTATION_HIGH);
-			this.iSecrecy = (Interface) reqPack.getPackagedElement(ANNOTATION_SECRECY);
-			this.iIntegrity = (Interface) reqPack.getPackagedElement(ANNOTATION_INTEGRITY);
-		} else {
-			throw new IllegalStateException("Couldn't find package containing UMLsec annotation types!");
-		}
 	}
 
 	public boolean processFwd() throws ProcessingException {
 		for (final Element e : this.model.allOwnedElements()) {
 			if (e instanceof Comment) {
-				final Comment comment = (Comment) e;
-				final String body = comment.getBody();
+				final var comment = (Comment) e;
+				final var body = comment.getBody();
 
-				final EList<Element> annotatedElements = comment.getAnnotatedElements();
+				final var annotatedElements = comment.getAnnotatedElements();
 				if (annotatedElements.isEmpty()) {
 					continue;
 				}
-				final Element element = annotatedElements.parallelStream()
+				final var element = annotatedElements.parallelStream()
 						.filter(el -> el.getOwnedComments().contains(comment)).findAny()
 						.orElse(annotatedElements.get(0));
 				if (ANNOTATION_CRITICAL.equals(body)) {
-					final critical critical = getCriticalStereotype((Classifier) element);
+					final var critical = getCriticalStereotype((Classifier) element);
 					if (!addValuesToCritical(comment, critical)) {
 						LOGGER.error("Couldnt add comment to <<critical>>");
 						return false;
 					}
 				} else {
-					final String tag = annotationNameToTagName(body);
+					final var tag = annotationNameToTagName(body);
 					if (tag == null) {
 						continue;
 					}
 
-					final critical critical = getCriticalStereotype(element);
-					final String signature = getSignature(element);
+					final var critical = getCriticalStereotype(element);
+					final var signature = getSignature(element);
 					if (!addValuesToCritical(critical, tag, signature)) {
 						LOGGER.error("Couldn't add " + tag + " = {\"" + signature + "\"} to <<criticial>>");
 						return false;
@@ -138,16 +109,63 @@ public class UmlSecProcessor {
 	 * @throws ProcessingException
 	 */
 	public boolean processBwd() throws ProcessingException {
+		if(!init(this.model)) {
+			// Nothing to do
+			return true;
+		}
 		Stream<EObject> delete = Stream.empty();
-
 		for (final Element element : this.model.allOwnedElements()) {
 			if ((element instanceof Classifier) && !(element instanceof PrimitiveType)) {
-				final Classifier classifier = (Classifier) element;
+				final var classifier = (Classifier) element;
 				delete = Stream.concat(delete, processBwd(classifier));
 			}
 		}
 		EMFUtil.deleteAll(delete.collect(Collectors.toSet()), this.model.eResource());
 		return true;
+	}
+
+	/**
+	 * Searches the UMLsec annotations in the model
+	 *
+	 * @param model A model of a program
+	 * @return
+	 */
+	private boolean init(final Model model) {
+		if(loaded()) {
+			return true;
+		}
+		final var namespace = new String[] { "org", "gravity", "security", "annotations", "requirements" };
+		final var stack = new LinkedList<Model>();
+		stack.add(model);
+		Package reqPack = null;
+		while (!stack.isEmpty()) {
+			final var next = stack.pop();
+			reqPack = getPackage(next, namespace);
+			if (reqPack != null) {
+				break;
+			}
+			if (next != null) {
+				for (final PackageableElement element : next.getPackagedElements()) {
+					if (element instanceof Model) {
+						stack.add((Model) element);
+					}
+				}
+			}
+		}
+		if (reqPack != null) {
+			this.iCritical = (Interface) reqPack.getPackagedElement(ANNOTATION_CRITICAL);
+			this.iHigh = (Interface) reqPack.getPackagedElement(ANNOTATION_HIGH);
+			this.iSecrecy = (Interface) reqPack.getPackagedElement(ANNOTATION_SECRECY);
+			this.iIntegrity = (Interface) reqPack.getPackagedElement(ANNOTATION_INTEGRITY);
+		}
+		return loaded();
+	}
+
+	/**
+	 * @return
+	 */
+	private boolean loaded() {
+		return (this.iCritical != null) && (this.iHigh != null) && (this.iSecrecy != null) && (this.iIntegrity != null);
 	}
 
 	/**
@@ -158,17 +176,17 @@ public class UmlSecProcessor {
 	 * @throws ProcessingException If the processing failed
 	 */
 	private Stream<Comment> processBwd(final Classifier classifier) throws ProcessingException {
-		final HashMap<String, Element> signatures = new HashMap<>();
-		final HashMap<String, Comment> highComments = new HashMap<>();
-		final HashMap<String, Comment> secrecyComments = new HashMap<>();
-		final HashMap<String, Comment> integrityComments = new HashMap<>();
+		final var signatures = new HashMap<String, Element>();
+		final var highComments = new HashMap<String, Comment>();
+		final var secrecyComments = new HashMap<String, Comment>();
+		final var integrityComments = new HashMap<String, Comment>();
 
-		final Comment criticalComment = getComments(classifier, signatures, highComments, secrecyComments,
+		final var criticalComment = getComments(classifier, signatures, highComments, secrecyComments,
 				integrityComments);
 
 		for (final EObject stereotype : classifier.getStereotypeApplications()) {
 			if (stereotype instanceof critical) {
-				final critical critical = (critical) stereotype;
+				final var critical = (critical) stereotype;
 
 				processBwd(classifier, signatures, highComments, criticalComment, critical.getHigh(), ANNOTATION_HIGH,
 						TAG_HIGH);
@@ -199,19 +217,19 @@ public class UmlSecProcessor {
 			criticalComment = getComment(classifier, ANNOTATION_CRITICAL, true);
 			criticalComment.getAnnotatedElements().add(this.iCritical);
 		}
-		final Comment tagComment = getComment(criticalComment, tagString, true);
+		final var tagComment = getComment(criticalComment, tagString, true);
 		for (final String value : tagValues) {
 			if (tagComments.remove(value) == null) {
 				// if the signature hasn't already a annotation
 				if (signatures.containsKey(value)) {
 					// if the classifier defines a member with the signature add a comment to
 					// the member
-					final Element member = signatures.get(value);
+					final var member = signatures.get(value);
 					createComment(memberAnnotationString, member);
 				} else {
 					// if the classifier doesn't defines a member with the signature add the
 					// signature to the according tag of the critical comment.
-					final Comment comment = UMLFactory.eINSTANCE.createComment();
+					final var comment = UMLFactory.eINSTANCE.createComment();
 					comment.setBody(value);
 					comment.getAnnotatedElements().add(tagComment);
 					tagComment.getOwnedComments().add(comment);
@@ -249,7 +267,7 @@ public class UmlSecProcessor {
 	 * @throws ProcessingException If the comment contains an unknown security level
 	 */
 	private void createComment(final String value, final Element member) throws ProcessingException {
-		final Comment comment = UMLFactory.eINSTANCE.createComment();
+		final var comment = UMLFactory.eINSTANCE.createComment();
 		comment.setBody(value);
 		comment.getAnnotatedElements().add(member);
 		if (High.class.getSimpleName().equals(value)) {
@@ -274,7 +292,7 @@ public class UmlSecProcessor {
 	private static Package getPackage(final Model model, final String... namespace) {
 		Package current = model;
 		for (final String element : namespace) {
-			final PackageableElement next = current.getPackagedElement(element);
+			final var next = current.getPackagedElement(element);
 			if ((next == null) || !next.eClass().equals(UMLPackage.eINSTANCE.getPackage())) {
 				return null;
 			}
@@ -294,8 +312,8 @@ public class UmlSecProcessor {
 			return false;
 		}
 		for (final Comment tag : comment.getOwnedComments()) {
-			final String tagName = tag.getBody();
-			final String[] signatures = tag.getOwnedComments().parallelStream().map(Comment::getBody)
+			final var tagName = tag.getBody();
+			final var signatures = tag.getOwnedComments().parallelStream().map(Comment::getBody)
 					.toArray(String[]::new);
 			addValuesToCritical(crit, tagName, signatures);
 		}
@@ -310,7 +328,7 @@ public class UmlSecProcessor {
 	 * @param signatures The signatures
 	 */
 	private boolean addValuesToCritical(final critical crit, final String tag, final String... signatures) {
-		final EStructuralFeature feature = crit.eClass().getEStructuralFeature(tag);
+		final var feature = crit.eClass().getEStructuralFeature(tag);
 		if (feature == null) {
 			LOGGER.error("Unknown <<critical>> tag: " + tag);
 			return false;
@@ -320,7 +338,7 @@ public class UmlSecProcessor {
 			return false;
 		}
 		@SuppressWarnings("unchecked")
-		final Collection<Object> values = (Collection<Object>) crit.eGet(feature);
+		final var values = (Collection<Object>) crit.eGet(feature);
 		Collections.addAll(values, signatures);
 		return true;
 	}
@@ -337,7 +355,7 @@ public class UmlSecProcessor {
 			getComments(signatures, highComments, secrecyComments, integrityComments, property);
 		}
 
-		final Comment criticalComment = getComment(classifier, ANNOTATION_CRITICAL, false);
+		final var criticalComment = getComment(classifier, ANNOTATION_CRITICAL, false);
 		if (criticalComment == null) {
 			return null;
 		}
@@ -363,10 +381,10 @@ public class UmlSecProcessor {
 	private static void getComments(final HashMap<String, Element> signatures,
 			final HashMap<String, Comment> highComments, final HashMap<String, Comment> secrecyComments,
 			final HashMap<String, Comment> integrityComments, final Element operation) {
-		final String signature = getSignature(operation);
+		final var signature = getSignature(operation);
 		signatures.put(signature, operation);
 		for (final Comment comment : operation.getOwnedComments()) {
-			final String body = comment.getBody();
+			final var body = comment.getBody();
 			if (High.class.getSimpleName().equals(body)) {
 				highComments.put(signature, comment);
 			} else if (Integrity.class.getSimpleName().equals(body)) {
@@ -426,7 +444,7 @@ public class UmlSecProcessor {
 				return (critical) eObject;
 			}
 		}
-		final critical critical = UmlsecFactory.eINSTANCE.createcritical();
+		final var critical = UmlsecFactory.eINSTANCE.createcritical();
 		critical.setBase_Classifier(classifier);
 		classifier.eResource().getContents().add(critical);
 		return critical;
